83- refactoring into folders, adding authentication to mongodb and adding a few comments
