+++ src/DatabaseToolkit/MongoDB/mongodb-backup-with-auth.bat
@@ -1,0 +1,1 @@
+mongodump -d %1 --gzip --archive=%2 --username %3 --authenticationDatabase %4
