+++ src/DatabaseToolkit/MongoDB/mongodb-restore-with-auth.bat
@@ -1,0 +1,1 @@
+mongorestore --drop --gzip --archive=%1 --username %2 --authenticationDatabase %3
