+++ src/DatabaseToolkit/MySQL/mysql-backup.bat
@@ -1,0 +1,1 @@
+%1 --defaults-file=%2 %3 > %4
