+++ src/DatabaseToolkit/appsettings.json
@@ -10,6 +10,9 @@
     "MySqlDefaultsFilePath": "C:\\ProgramData\\MySQL\\MySQL Server 8.0\\my.ini",
     "PostgreSQLHost": "localhost",
     "PostgreSQLPort": "5432",
-    "PostgreSQLUser": "postgres"
+    "PostgreSQLUser": "postgres",
+    "MongoDBUser": "admin",
+    "MongoDBPassword": "password",
+    "MongoDBAuthenticationDatabase": "admin"
   }
 }
